requirements.txt
# Food Scanner - Dependencies
google-generativeai>=0.8.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
Pillow>=10.0.0
streamlit>=1.30.0
